+++ src/commands/config.rs
@@ -22,9 +22,9 @@ pub struct Args {
 }
 
 impl Args {
-    /// Returns `true` if the subcommand is 'Init'.
-    pub fn is_init(&self) -> bool {
-        matches!(self.command, ConfigSubcommand::Init)
+    /// Returns `true` if the subcommand is 'Default'.
+    pub fn is_default(&self) -> bool {
+        matches!(self.command, ConfigSubcommand::Default)
     }
 }
 
@@ -35,7 +35,7 @@ pub enum ConfigSubcommand {
     Schema,
 
     /// Generates a default configuration file.
-    Init,
+    Default,
 
     /// Displays the current configuration.
     Resolve(ResolveArgs),
@@ -60,7 +60,7 @@ pub fn config(args: Args, mut config: Config) -> CommandResult<()> {
             println!("{schema_pretty}");
             return Ok(());
         }
-        ConfigSubcommand::Init => {
+        ConfigSubcommand::Default => {
             include_schema_directive = true;
             Config::default()
         }